Fix UTF-8 kernel availability when utf8proc is unavailable
